src/cost_monitoring/requirements.txt
Flask==2.2.2
google-cloud-billing==1.9.0
google-cloud-service-usage==1.4.0
google-cloud-pubsub==2.13.4
google-cloud-firestore==2.7.2
PyYAML==6.0
